# src/main/java/com/projectx/model/Service.java
package com.projectx.model;

import java.util.Date;

/**
 * Created by ivan on 13.04.17.
 */
public class Service {
    private long id;
    private String name;
    private String description;
    private int rating;
    private int price;
    private long userId;
    private Date dateCreated;

    public Service() {}

    public Service(String name, String description, int rating, int price, long userId) {
        this.name = name;
        this.description = description;
        this.rating = rating;
        this.price = price;
        this.userId = userId;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public int getRating() {
        return rating;
    }

    public void setRating(int rating) {
        this.rating = rating;
    }

    public int getPrice() {
        return price;
    }

    public void setPrice(int price) {
        this.price = price;
    }

    public long getUserId() {
        return userId;
    }

    public void setUserId(long userId) {
        this.userId = userId;
    }

    public Date getDateCreated() {
        return dateCreated;
    }

    public void setDateCreated(Date dateCreated) {
        this.dateCreated = dateCreated;
    }
}